docker_example/requirements.txt
scipy
nibabel
torchio
monai[nibabel]==1.1.0
tensorboard
wandb
pandas
numba
lpips==0.1.4
einops>=0.3.2
omegaconf==2.1.1
imageio
seaborn
scikit-image
x-transformers==1.8.1
pytorch-lightning
